requirements.txt
django
djangorestframework
Pillow
crispy-bootstrap5
django-crispy-forms
requests
djangorestframework-simplejwt
python-dotenv
